requirements.txt
streamlit==1.51.0
google-generativeai==0.8.4
python-dotenv==1.0.1
flask==3.0.3
flask-cors==4.0.1
